backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
opencv-python>=4.9.0
mediapipe>=0.10.14
scikit-learn>=1.4.0
numpy>=1.26.0
pyttsx3>=2.90
pillow>=10.3.0
python-dotenv>=1.0.1
httpx>=0.27.0       # used by test client
pytest>=8.2.0
pytest-asyncio>=0.23.0

# ── MongoDB ─────────────────────────────────────────────────────────
motor>=3.3.0
pymongo>=4.6.0

# ── Auth ──────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1          # passlib is incompatible with bcrypt>=4.1
python-multipart>=0.0.9
email-validator>=2.2.0
